refactor: correctly path for infrastructure dir
