+++ app/.env
@@ -1,0 +1,2 @@
+PAYMENT_PROCESSOR_URL_DEFAULT=http://localhost:8001
+PAYMENT_PROCESSOR_URL_FALLBACK=http://localhost:8002
