+++ app/.env.sample
@@ -1,0 +1,2 @@
+PAYMENT_PROCESSOR_URL_DEFAULT=http://payment-processor-default:8080
+PAYMENT_PROCESSOR_URL_FALLBACK=http://payment-processor-fallback:8080
