+++ app/build.gradle.kts
@@ -26,6 +26,7 @@ dependencies {
     implementation(libs.guava)
     implementation(libs.javalin)
     implementation(libs.slf4j.simple)
+    implementation(libs.dotenv.java)
 }
 
 // Apply a specific Java toolchain to ease working on different environments.
