+++ gradle/libs.versions.toml
@@ -6,9 +6,11 @@ guava = "33.4.5-jre"
 junit-jupiter = "5.12.1"
 javalin = "6.7.0"
 slf4j-simple = "2.0.16"
+dotenv-java = "3.2.0"
 
 [libraries]
 guava = { module = "com.google.guava:guava", version.ref = "guava" }
 junit-jupiter = { module = "org.junit.jupiter:junit-jupiter", version.ref = "junit-jupiter" }
 javalin = { module = "io.javalin:javalin", version.ref = "javalin" }
 slf4j-simple = { module = "org.slf4j:slf4j-simple", version.ref = "slf4j-simple" }
+dotenv-java = { module = "io.github.cdimascio:dotenv-java", version.ref = "dotenv-java" }
